chore: Update pr.ts to add comment and delete existing comments
This commit updates the pr.ts file to add a new function `addCommentToPR` that adds a comment to a pull request. It also adds a new function `deleteExistingComments` that deletes existing comments from a pull request. These changes are made to improve the functionality of the pull request review process.