feat: Add function to retrieve system variables
This commit adds a new function `getSystemVariables` to the `pr.ts` file. This function retrieves system variables from the runtime environment, which are typically set by the CI/CD system. The system variables include the URI of the system team collection, the ID of the system project, the name of the build repository, the ID of the system pull request, the system access token, and the system project. This addition improves the functionality of the pull request review process.